requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
joblib>=1.3.0
matplotlib>=3.7.0
streamlit>=1.25.0
